Implement `which-update`
