+++ Library/Homebrew/sorbet/rbi/dsl/homebrew/dev_cmd/which_update.rbi
@@ -1,0 +1,34 @@
+# typed: true
+
+# DO NOT EDIT MANUALLY
+# This is an autogenerated file for dynamic methods in `Homebrew::DevCmd::WhichUpdate`.
+# Please instead update this file by running `bin/tapioca dsl Homebrew::DevCmd::WhichUpdate`.
+
+
+class Homebrew::DevCmd::WhichUpdate
+  sig { returns(Homebrew::DevCmd::WhichUpdate::Args) }
+  def args; end
+end
+
+class Homebrew::DevCmd::WhichUpdate::Args < Homebrew::CLI::Args
+  sig { returns(T::Boolean) }
+  def commit?; end
+
+  sig { returns(T::Boolean) }
+  def eval_all?; end
+
+  sig { returns(T::Boolean) }
+  def install_missing?; end
+
+  sig { returns(T.nilable(String)) }
+  def max_downloads; end
+
+  sig { returns(T::Boolean) }
+  def stats?; end
+
+  sig { returns(T.nilable(String)) }
+  def summary_file; end
+
+  sig { returns(T::Boolean) }
+  def update_existing?; end
+end
