+++ Library/Homebrew/test/dev-cmd/which-update_spec.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require "cmd/shared_examples/args_parse"
+require "dev-cmd/which-update"
+
+RSpec.describe Homebrew::DevCmd::WhichUpdate do
+  it_behaves_like "parseable arguments"
+end
